+++ examples/deps.cmake
@@ -1,7 +1,3 @@
 set(APP_DEPENDENT_MODULES "lib_xua"
-                          "lib_xud(2.3.2)"
-                          "lib_i2c(6.2.0)"
-                          "lib_spdif(6.1.1)"
-                          "lib_mic_array(4.6.0)"
                           "lib_board_support(1.0.1)"
                           )